fastapi/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
qrcode>=7.4.2
Pillow>=10.2.0
python-multipart>=0.0.6
websockets>=12.0
